Restructuring completed
