+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/travis_logo"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.MainActivity"
+        <activity android:name=".views.activities.MainActivity"
             android:theme="@style/MainTheme"/>
-        <activity android:name=".activities.SplashScreenActivity"
+        <activity android:name=".views.activities.SplashScreenActivity"
             android:theme="@style/MainTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
