+++ app/src/main/java/com/aad_team_42/travelmanticsrebranded/views/activities/SplashScreenActivity.java
@@ -1,11 +1,10 @@
-package com.aad_team_42.travelmanticsrebranded.activities;
+package com.aad_team_42.travelmanticsrebranded.views.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.widget.Toast;
 
 import com.aad_team_42.travelmanticsrebranded.R;
