+++ app/src/main/java/com/aad_team_42/travelmanticsrebranded/views/fragments/ExploreFragment.java
@@ -1,4 +1,4 @@
-package com.aad_team_42.travelmanticsrebranded.activities.fragments;
+package com.aad_team_42.travelmanticsrebranded.views.fragments;
 
 
 import android.os.Bundle;
@@ -10,11 +10,10 @@ import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.LinearLayout;
 
 import com.aad_team_42.travelmanticsrebranded.R;
-import com.aad_team_42.travelmanticsrebranded.activities.adapters.ExploreAdapter;
-import com.aad_team_42.travelmanticsrebranded.activities.model.Explore;
+import com.aad_team_42.travelmanticsrebranded.adapters.ExploreAdapter;
+import com.aad_team_42.travelmanticsrebranded.model.Explore;
 
 import java.util.ArrayList;
 import java.util.List;
